requirements.txt
spacy
gensim
pytorch-crf
label-studio
torchtext==0.6.0